style: 首页0.5
